Move database functions to ChannelQueryes file
